+++ desktop/packages/mullvad-vpn/src/config.json
@@ -1,10 +1,1 @@
-{
-  "supportEmail": "support@mullvadvpn.net",
-  "strings": {
-    "wireguard": "WireGuard",
-    "openvpn": "OpenVPN",
-    "splitTunneling": "Split tunneling",
-    "daita": "DAITA",
-    "daitaFull": "Defence against AI-guided Traffic Analysis"
-  }
-}
+{}
