+++ desktop/packages/mullvad-vpn/src/renderer/components/select-location/SelectLocation.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useState } from 'react';
 import { sprintf } from 'sprintf-js';
 
-import { strings } from '../../../config.json';
+import { strings } from '../../../shared/constants';
 import { Ownership } from '../../../shared/daemon-rpc-types';
 import { messages } from '../../../shared/gettext';
 import { IconButton } from '../../lib/components';
