+++ desktop/packages/mullvad-vpn/src/shared/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './links';
+export * from './strings';
