+++ desktop/packages/mullvad-vpn/src/shared/constants/strings.ts
@@ -1,0 +1,8 @@
+export const strings = {
+  wireguard: 'WireGuard',
+  openvpn: 'OpenVPN',
+  splitTunneling: 'Split tunneling',
+  daita: 'DAITA',
+  daitaFull: 'Defence against AI-guided Traffic Analysis',
+  supportEmail: 'support@mullvadvpn.net',
+} as const;
